# improve GetRegistrationResult